Let's put the kibosh on --namespace, too
This will result in the default behavior of installing into
knative-eventing, no matter where the Install lives.

And we optionally override the ns in the spec instead of a flag now.

Also added a predicate to the watch to ignore events as a result of
updating our Install's status.

Plus OLM requires a different downward api to support AllNamespaces
mode